Multiple Events Support
When creating a talk an exising event needs to be selected or can be created a new. Once published the talk will be listed for that event. A new event index lists all public events. Events can also be cerated with a password. In this case one has to now either the url or the event slug and the password to view the published talks of that event.